refactor(deck_notes_panel): split DeckNotesPanel into mixin package
Move widgets/panels/deck_notes_panel.py into a package: frame.py keeps
the note-type constants, the _NoteCardWidget subpanel, and the main
_build_ui; properties.py exposes the translate helper and get_notes;
handlers.py owns set_notes/clear/load/save, the card-widget rebuild,
and the event callbacks.
