+++ widgets/panels/deck_notes_panel/__init__.py
@@ -1,0 +1,7 @@
+"""Deck notes panel package."""
+
+from __future__ import annotations
+
+from widgets.panels.deck_notes_panel.frame import DeckNotesPanel
+
+__all__ = ["DeckNotesPanel"]
